app/requirements.txt
click==8.1.3
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Login==0.6.2
flask-mongoengine==1.0.0
Flask-WTF==1.1.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mongoengine==0.27.0
pymongo==4.3.3
six==1.14.0
Werkzeug==2.2.2
WTForms==2.3.3
python-dotenv==0.19.1